# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/products/add": {
            "post": {
                "description": "Save new product into DynamoDb",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a Product item",
                "parameters": [
                    {
                        "description": "Product item to add",
                        "name": "entityBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.ProductBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/entities.ProductResponse"
                        }
                    }
                }
            }
        },
        "/products/delete/{id}": {
            "delete": {
                "description": "Save new product into DynamoDb",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a Product item by Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the product to find",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/entities.ProductResponse"
                        }
                    }
                }
            }
        },
        "/products/find/{id}": {
            "get": {
                "description": "Select new product into DynamoDb",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Find a Product item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the product to find",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/entities.ProductResponse"
                        }
                    }
                }
            }
        },
        "/products/list": {
            "get": {
                "description": "Select all products from DynamoDb",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/entities.ProductListResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.ProductBody": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "entities.ProductListResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.ProductBody"
                    }
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "entities.ProductResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}